Refactored backend structure: Split stone and pricing controllers into separate files (controllers, services, utils) and improved modularity across backend files.
