+++ backend/src/routes/stones.js
@@ -8,10 +8,10 @@ const {
 } = require("../controllers/stoneController");
 
 const router = express.Router();
-router.get("/recommendations/:id", getRecommendedStones);
 router.get("/search", searchStones);
-router.get("/", getStones);
+router.get("/recommendations/:id", getRecommendedStones);
 router.get("/:id", getStoneById);
+router.get("/", getStones);
 router.post("/", createStone);
 
 module.exports = router;
